Moved zig, - In own file in prep for move to own module - Tests updated to import the module as zig wont be a default - Ensured tests working as is with just module change - Cleared any uneeded refs in seneca.js
